+++ src/gwh.ts
@@ -10,13 +10,13 @@ program
     'Migrate the database with the last schema of graphql-web-hook. We advise to provide a separated schema.'
   )
   .action(async function (configPath) {
-    let config = null
+    let dbConfig = null
     try {
-      config = require(configPath)
+      dbConfig = require(configPath)
     } catch (e: any) {
       throw new Error('Could not load the given config.' + e.message)
     }
-    const models = getModels(config)
+    const models = getModels({dbConfig})
     await migrate(models)
     await models.sequelize.close()
   })